refactor: extract types and utilities from App.vue
- src/types.ts: shared type definitions (Task, WeekData, DayKey, etc.)
  and constants (ALL_KEYS, DAY_LABELS, etc.)
- src/utils/dateUtils.ts: pure date functions (getISOWeek, getWeekMonday,
  getWeekDates, toDateStr)
- src/utils/weekData.ts: week data helpers (emptyWeek, normalizeWeekData)

Reduces App.vue from 1716 to 1601 lines with no behavioral changes.
